pyproject.toml
[project]
name = "npi-mcp"
version = "0.1.0"
description = "MCP Server for the NPI registry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0b2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ty>=0.0.15",
    "vcrpy>=8.1.1",
]
